python_api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets>=12.0
pymysql==1.1.1
python-dotenv==1.0.1
pydantic==2.7.1
httpx==0.27.0

# ML stack
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2

# Email
requests==2.31.0
